[Ubuntu 18.04] General changes to AppliedRoboticsEnvironment
> Changed CMakeLists.txt to use OpenCV 3
> Commented out 'DEPENDS system_lib' since system_lib is not a real library
> Changed gazebo_ros_interface.cpp to use ignition::math instead of gazebo::math
> Updated READMe.md requirements and dependencies
> Changed CMakeLists.txt to use Gazebo 9 instead of Gazebo 7

+++ src/4_planning/include/planning_handle.hpp
@@ -51,7 +51,7 @@ private:
   //std::vector<std::pair<int,Polygon>> victim_list_;
   Polygon gate_, borders_;
   std::vector<float> x_, y_, theta_;
-  double x0_, x1_, x2_, y0_, y1_, y2_, theta0_, theta1_, theta2_;
+  float x0_, x1_, x2_, y0_, y1_, y2_, theta0_, theta1_, theta2_;
   std::vector<Path> path_;
 
   std_msgs::Header header_;
